Better fallbacks + name consistency
