+++ package.json
@@ -10,7 +10,7 @@
   },
   "dependencies": {
     "@react-three/fiber": "^7.0.1",
-    "next": "10.2.3",
+    "next": "^10.2.3",
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "styled-components": "^5.2.3",
@@ -19,7 +19,7 @@
   },
   "devDependencies": {
     "@types/react": "17.0.11",
-    "@types/styled-components": "^5.2.3",
+    "@types/styled-components": "^5.1.15",
     "@types/three": "^0.129.1",
     "eslint": "7.28.0",
     "eslint-config-next": "11.0.0",
